refactor : useBookmarks 추가 (#60)
