+++ src/components/place_detail/PlaceDetail.tsx
@@ -1,7 +1,7 @@
-import { deleteBookmark, getBookmark, insertBookmark } from '@/apis/bookmarks';
+import { getBookmark } from '@/apis/bookmarks';
 import { RootState } from '@/redux/config/configStore';
 import { Tables } from '@/types/supabase';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import { FaRegBookmark, FaBookmark } from 'react-icons/fa';
 import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
@@ -9,6 +9,7 @@ import { toastSuccess, toastWarn } from '@/libs/toastifyAlert';
 import { Chip } from '@nextui-org/react';
 import { RiKakaoTalkFill } from 'react-icons/ri';
 import { shareKakao } from '@/utils/shareKaKao';
+import { useBookmarks } from '@/hooks/useBookmarks';
 
 interface PlaceInfoAllData {
   placeId: string;
@@ -16,10 +17,10 @@ interface PlaceInfoAllData {
 }
 
 const PlaceDetail = ({ placeInfo, placeId }: PlaceInfoAllData) => {
-  const queryClient = useQueryClient();
   const [isBookmarked, setIsBookmarked] = useState<boolean>(false);
   const { userId, isLoggedIn } = useSelector((state: RootState) => state.auth);
   const { place_name, tel, address, working_hours, holidays } = placeInfo;
+  const { insertBookmark, deleteBookmark } = useBookmarks(userId, placeId);
 
   const { data: bookmarkState } = useQuery({
     queryKey: ['bookmark', userId, placeId],
@@ -53,63 +54,15 @@ const PlaceDetail = ({ placeInfo, placeId }: PlaceInfoAllData) => {
     '휠체어 대여',
   ];
 
-  // 낙관적 업데이트 (추가)
-  const { mutate: addBookmark } = useMutation({
-    mutationFn: insertBookmark,
-    onMutate: async () => {
-      await queryClient.cancelQueries({
-        queryKey: ['bookmark', userId, placeId],
-      });
-      const prev = queryClient.getQueryData(['bookmark', userId, placeId]);
-      const updateBookmark = [{ userId, place_id: placeId }];
-      queryClient.setQueryData(['bookmark', userId, placeId], updateBookmark);
-      return { prev };
-    },
-    onError: (error, updateReviewParams, context) => {
-      if (context?.prev) {
-        queryClient.setQueryData(['bookmark', userId, placeId], context.prev);
-      }
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({
-        queryKey: ['bookmark', userId, placeId],
-      });
-    },
-  });
-
-  // 낙관적 업데이트 (삭제)
-  const { mutate: delBookmark } = useMutation({
-    mutationFn: deleteBookmark,
-    onMutate: async () => {
-      await queryClient.cancelQueries({
-        queryKey: ['bookmark', userId, placeId],
-      });
-      const prev = queryClient.getQueryData(['bookmark', userId, placeId]);
-      const updateBookmark = undefined;
-      queryClient.setQueryData(['bookmark', userId, placeId], updateBookmark);
-      return { prev };
-    },
-    onError: (error, updateReviewParams, context) => {
-      if (context?.prev) {
-        queryClient.setQueryData(['bookmark', userId, placeId], context.prev);
-      }
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({
-        queryKey: ['bookmark', userId, placeId],
-      });
-    },
-  });
-
   // 버튼 토글
   const toggleBookmark = () => {
     if (isBookmarked) {
       setIsBookmarked(false);
-      delBookmark({ userId, placeId });
-      toastSuccess('북마크에 해제되었습니다');
+      deleteBookmark({ userId, placeId });
+      toastSuccess('북마크가 해제되었습니다');
     } else {
       setIsBookmarked(true);
-      addBookmark({ userId, placeId });
+      insertBookmark({ userId, placeId });
       toastSuccess('북마크에 추가되었습니다');
     }
   };
